requirements.txt
flask~=3.1.0
flask-login~=0.6.3
flask-mysql
mysqlclient
mysql-connector
tinymongo~=0.2.0
tinydb~=4.8.2
Flask-SQLAlchemy
flask-migrate~=4.0.7
python-dotenv~=1.0.1
Werkzeug~=3.1.3
alembic~=1.14.0
selenium